+++ Sources/MasKit/Models/SearchResult.swift
@@ -15,7 +15,7 @@ struct SearchResult: Decodable {
     var price: Double?
     var sellerName: String
     var sellerUrl: String?
-    var trackId: Int
+    var trackId: UInt64
     var trackName: String
     var trackViewUrl: String
     var version: String
@@ -29,7 +29,7 @@ struct SearchResult: Decodable {
         price: Double = 0.0,
         sellerName: String = "",
         sellerUrl: String = "",
-        trackId: Int = 0,
+        trackId: UInt64 = 0,
         trackName: String = "",
         trackViewUrl: String = "",
         version: String = ""
